% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List Available Palettes}
\usage{
list_palettes()
}
\value{
A character vector of palette names
}
\description{
Display all available color palettes in the package.
}
\examples{
list_palettes()
}
