% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes_list}
\alias{palettes_list}
\title{Color Palettes Data}
\format{
A list of named character vectors containing hex color codes
}
\usage{
palettes_list
}
\description{
A list containing all available color palettes in the package.
Each palette is a named vector of hex color codes.
}
\keyword{datasets}
